requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40
fastapi>=0.100.0
uvicorn>=0.20.0
starlette>=0.27.0

# LLM Providers
openai>=1.0.0
langchain-google-genai>=0.0.5

# Database and Vector Store
# Note: sqlite3 is built-in with Python
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
langchain-huggingface>=0.0.1

# Document Processing
pypdf>=3.0.0
python-multipart>=0.0.6
# pydf2>=3.0.0
# reportlab>=3.6.12 check fir this
# Configuration and Data
pyyaml>=6.0
pydantic>=2.0.0
jinja2>=3.1.0

# HTTP and Communication
httpx>=0.24.0
aiofiles>=23.0.0
websockets>=11.0.0

# Streamlit for UI
streamlit>=1.28.0

# Optional: Monitoring and Observability (comment out if not needed)
# langfuse>=2.0.0

# Development and Testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Optional: Monitoring (only install if you want observability)
# langsmith>=0.0.87

# Environment Management
python-dotenv>=1.0.0